+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # EPS Release History
 
+## 0.5.1
+
+ * Removed dead code (could not be called) [Dominique Broeglin]
+ * Fixed #25 and other edge cases around literal `<%%` and `%%>` [Dominique Broeglin]
+
+## 0.5.0
+
+ * Allow Get-OrElse to use pipeline input [Dominique Broeglin]
+ * Improved README documentation [Dominique Broeglin]
+ * Added test in context of #20 [Dominique Broeglin]
+ * Added Get-OrElse function (optimized function loading when using -Safe) [Dominique Broeglin]
+ * Better unit testing of the 'Each' helper function [Dominique Broeglin]
+ * Allow relative path to 'Invoke-EpsTemplate -Path' [kimura wataru]
+
+## 0.4.0
+
+ * Added an 'Each' function for iteration joining [Dominique Broeglin]
+
 ## 0.3.3 (2017-03-11)
 
  * Also fixed and issue when concatenating two code blocks [Dominique Broeglin]